DSND-3007: Create base class for versioned and unversioned metrics document
