+++ src/test/java/uk/gov/companieshouse/company/metrics/service/CompanyMetricsServiceTest.java
@@ -12,7 +12,6 @@ import static org.mockito.Mockito.when;
 
 import java.io.IOException;
 import java.time.OffsetDateTime;
-import java.time.format.DateTimeFormatter;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -21,7 +20,13 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.companieshouse.api.metrics.*;
+import uk.gov.companieshouse.api.metrics.AppointmentsApi;
+import uk.gov.companieshouse.api.metrics.MetricsApi;
+import uk.gov.companieshouse.api.metrics.MetricsRecalculateApi;
+import uk.gov.companieshouse.api.metrics.MortgageApi;
+import uk.gov.companieshouse.api.metrics.PscApi;
+import uk.gov.companieshouse.api.metrics.RegisterApi;
+import uk.gov.companieshouse.api.metrics.RegistersApi;
 import uk.gov.companieshouse.company.metrics.model.CompanyMetricsDocument;
 import uk.gov.companieshouse.company.metrics.model.TestData;
 import uk.gov.companieshouse.company.metrics.model.Updated;
@@ -111,7 +116,8 @@ class CompanyMetricsServiceTest {
                 createUpdated("source-metrics-updated-body-1.json");
 
         CompanyMetricsDocument companyMetricsDocument = new CompanyMetricsDocument(metricsApi,
-                updated);
+                updated)
+                .version(1L);
 
         when(companyMetricsRepository.findById(anyString()))
                 .thenReturn(Optional.of(companyMetricsDocument));
